+++ src/plugins/widgets/ipInfo/IpInfo.tsx
@@ -10,21 +10,37 @@ const IpInfo: React.FC<Props> = ({
   loader,
 }) => {
   const pushError = usePushError();
-  React.useEffect(() => {
+
+  const refreshData = React.useCallback(() => {
     getIpInfo(loader).then(setCache).catch(pushError);
-  }, []);
+  }, [loader, setCache, pushError]);
+
+  React.useEffect(() => {
+    refreshData();
+  }, [refreshData]);
 
   if (!cache) {
     return null;
   }
 
-  const ip = data.maskIP ? cache.ip.split('.').map((s, i) => i > 0 && i < 3 ? s.replace(/[\d]+/, '*') : s).join('.') : cache.ip;
+  const ip = data.maskIP 
+    ? cache.ip.split('.').map((s, i) => i > 0 && i < 3 ? s.replace(/[\d]+/, '*') : s).join('.') 
+    : cache.ip;
   const info = [];
   if (!data.hideIP) info.push(ip);
   if (data.displayCity) info.push(cache.city);
   if (data.displayCountry) info.push(cache.country);
 
-  return <div className="IpInfo">{info.join(", ")}</div>;
+  return (
+    <div 
+      className="IpInfo"
+      onClick={data.clickToRefresh ? refreshData : undefined}
+      style={{ cursor: data.clickToRefresh ? 'pointer' : 'default' }}
+      title={data.clickToRefresh ? "Click to refresh IP info" : undefined}
+    >
+      {info.join(", ")}
+    </div>
+  );
 };
 
 export default IpInfo;